+++ models/Cart.js
@@ -1,29 +1,12 @@
-const mongoose = require('mongoose')
+///////////////////////////////////////////////////////////////////////////////
+// User Cart
+///////////////////////////////////////////////////////////////////////////////
+const userCart = [];
 
-const CartSchema = new mongoose.Schema(
-  {
-    userId: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: "User"
-    },
-    products: [
-      {
-        productId: Number,
-        name: String,
-        price: Number,
-        quantity: Number,
-      }
-    ],
-    active: {
-      type: Boolean,
-      default: true
-    },
-    modifiedOn: {
-      type: Date,
-      default: Date.now
-    }
-  },
-  { timestamps: true }
-);
 
-module.exports = mongoose.model("Cart", CartSchema);
+function getNumUserItems() {
+  if (userCart.length === 0) return 0;
+  return userCart.reduce((acc, cur) => acc + cur.quantity, 0);
+}
+
+module.exports = { userCart, getNumUserItems };
