+++ views/index.ejs
@@ -13,12 +13,11 @@
         <!--bootStrap-->
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
         <!-- custom css file link  -->
-        <link rel="stylesheet" href="css/style.css">
+        <link rel="stylesheet" href="/public/css/style.css">
 
     </head>
     <body>
         <%- include('partials/header'); %>  
-        <!-- TODO: numCartItems not working  -->
 
     <!-- Deals Section -->
 
@@ -51,7 +50,7 @@
                 </div>
 
                 <div class="box">
-                    <img src="images/computers.png" alt="">
+                    <img src="images/macAd.jpeg" alt="">
                     <div class="content">
                         <span>upto 50% off</span>
                         <h3>PC & Laptops</h3>
@@ -67,10 +66,10 @@
                     </div>
                 </div>
                 <div class="box">
-                    <img src="images/cat-2.jpg" alt="">
+                    <img src="images/component.jpg" alt="">
                     <div class="content">
-                        <span>upto 50% off</span>
-                        <h3>for mans</h3>
+                        <span>up to 50% off</span>
+                        <h3>COMPONENTS & STORAGE</h3>
                         <a href="/products" class="btn">shop now</a>
                     </div>
                 </div>
@@ -104,31 +103,13 @@
                     </div>
                     <p><%= products[0].description %></p>
                     <div class="price">$<%= (products[0].default_price.unit_amount/100).toFixed(2) %></div>
-                    <a class="btn btn-add-to-cart btn-pill" data-product-id="<%= products[0].id %>">add to cart</a>
+                    <a id = "button" class="btn btn-add-to-cart" data-product-id="<%= products[0].id %>">add to cart</a>
                 </div>
             </div>
 
             <div class="box">
-                <!-- <section class="products">
-                    <% products.forEach(product => { %>
-                        <div class="product">
-                            <div class="product-top">
-                                <h2 class="product-name"><%= product.name %></h2>
-                                <img class="product-image" src="<%= product.images[0] || 'https://placehold.co/600x400?text=PRODUCT%20IMAGE' %>" />
-                            </div>
-                            <div>
-                                <p class="product-price">$<%= (product.default_price.unit_amount/100).toFixed(2) %></p>
-                                <button class="btn-add-to-cart btn-pill" data-product-id="<%= product.id %>">Add To Cart</button>
-                            </div>
-                        </div>
-                    <% }); %>
-                </section> -->
-
                 <div class="image-container">
                     <div class="small-image">
-                        <!-- <img src="images/f-2-1.png" class="small-image-2" alt="">
-                        <img src="images/f-2-2.png" class="small-image-2" alt="">
-                        <img src="images/f-2-3.png" class="small-image-2" alt=""> -->
                     </div>
                     <div class="big-image">
                         <img class="product-image" src="<%= products[3].images[0] || 'https://placehold.co/600x400?text=PRODUCT%20IMAGE' %>" />
@@ -146,7 +127,7 @@
                     </div>
                     <p><%= products[3].description %></p>
                     <div class="price">$<%= (products[3].default_price.unit_amount/100).toFixed(2) %></div>
-                    <a class="btn btn-add-to-cart btn-pill" data-product-id="<%= products[3].id %>">add to cart</a>
+                    <a id = "button" class="btn btn-add-to-cart" data-product-id="<%= products[3].id %>">add to cart</a>
                 </div>
             </div>
 
@@ -173,7 +154,7 @@
                     </div>
                     <p><%= products[1].description %></p>
                     <div class="price">$<%= (products[1].default_price.unit_amount/100).toFixed(2) %></div>
-                    <a class="btn btn-add-to-cart btn-pill" data-product-id="<%= products[1].id %>">add to cart</a>
+                    <a id = "button" class="btn btn-add-to-cart" data-product-id="<%= products[1].id %>">add to cart</a>
                 </div>
             </div>
