+++ views/partials/header.ejs
@@ -31,7 +31,7 @@
 
     <div class="user">
         <img src="images/user-img.png" alt="">
-        <h3>shaikh anas</h3>
+        <h3></h3>
     </div>
 
     <div class="links">
