+++ views/products.ejs
@@ -4,13 +4,20 @@
 	<meta charset="UTF-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Simple Ecommerce Sample</title>
-	<link href="public/css/reset.css" rel="stylesheet">
-	<link href="public/css/cart_styles.css" rel="stylesheet">
-	<link rel="stylesheet" href="public/css/style.css">
+	<title>Products - Tectonic</title>
+	<link href="/public/css/reset.css" rel="stylesheet">
+	<link href="/public/css/cart_styles.css" rel="stylesheet">
+	<link rel="stylesheet" href="/public/css/style.css">
+	<link rel="stylesheet" href="https://unpkg.com/swiper/swiper-bundle.min.css" />
+	<!-- font awesome cdn link  -->
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
+        <!--bootStrap-->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+        
 </head>
 <body>
 	<%- include('partials/header', { numCartItems }) %>
+
 	<main>
 		<div class="container">
 			<section class="products">
